Notifications
Basic notifications feature.

Does NOT check if the room of the notification is currently open to skip
showing. Thus, the feature is very annoying in its current state for most users.

Can be enabled in config.json, disabled by default because of the above
missing check.

There are no checks against sizes of message/body, usernames, or room
names.