Users who have the same initials will now have different coloured read receipts (until all colours are used, then it loops).
This is implemented by assigning a `Color` to each user who has a valid
read recipt in a room, and stored as a `HashMap<OwnedUserId, Color>` in
the `RoomInfo` struct.

Overlaps (up until all colors used) are prevented by checking how many
exisiting users have the same inital, then adding that number to the
colour index of the 'first' user in the map with the same initial.